requirements.txt
streamlit>=1.42.0
streamlit-authenticator>=0.4.1
anthropic>=0.50.0
pdfplumber>=0.11.0
PyMuPDF>=1.25.0
pymupdf4llm>=0.0.17
python-docx>=1.1.0
matplotlib>=3.9.0
plotly>=5.24.0
pandas>=2.2.0
pydantic>=2.10.0
openpyxl>=3.1.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pyyaml>=6.0
openai>=1.30.0
python-dotenv>=1.0.0
jinja2>=3.1.0
kaleido>=0.2.1
bcrypt>=4.0.0
weasyprint>=60.0
pytest>=8.0.0
pytest-cov>=5.0.0

# Optional: for scanned PDF support (adds ~5GB with PyTorch + model weights)
# pip install surya-ocr
# surya-ocr>=0.17.0
